✅ 🚑 Change length estimate to only use _important_ characters.
Only characters that are classified as a major character should be used in length calculation. Tests and method updated accordingly.

Fixes #10

✅ Add additional tests

Add additional tests for forms. the missing tests left, when I compare them to the test code actually **ARE** being tested, but for some reason coverage is not catching that the scenario is being checked.

Fixes 11